fix(jobalert): plain-Node .ts import + circular import (PR review)
- provinceCantonAffinity.ts imported '../data/municipalities' without
  extension; Vite/vitest resolve it but plain Node (used by the daily
  send-job-alerts.mjs cron) throws ERR_MODULE_NOT_FOUND. Verified fix
  with `node -e "import('./services/jobAlertMatching.mjs')"`.
- Extracted SECTORS into jobAlertConstants.ts to break the
  JobAlertForm <-> ProfileEnrichmentPrompt circular import.
