+++ components/community/JobAlertForm.tsx
@@ -15,6 +15,7 @@ import { ABOVE_MOBILE_NAV_BOTTOM } from '@/components/shared/mobileNavClearance'
 import { consumeJobAlertOpen } from '@/services/jobAlertOpenSignal';
 import { savePendingJobAlert, consumePendingJobAlert } from '@/services/pendingJobAlert';
 import ProfileEnrichmentPrompt from './ProfileEnrichmentPrompt';
+import { SECTORS } from './jobAlertConstants';
 import { loadEnrichmentProfileFields } from '@/services/profileFirestore';
 import {
   loadGatingState,
@@ -52,19 +53,6 @@ const CONTRACT_TYPES = [
  { value: 'part-time', labelKey: 'jobBoard.contract.partTime' },
  { value: 'temporary', labelKey: 'jobBoard.contract.temporary' },
  { value: 'internship', labelKey: 'jobBoard.contract.internship' },
-];
-
-export const SECTORS = [
- { value: 'Fintech / Blockchain', label: 'Fintech' },
- { value: 'Tecnologia / Data Center', label: 'Tecnologia' },
- { value: 'Consulenza', label: 'Consulenza' },
- { value: 'Sanità / Assistenza', label: 'Sanità' },
- { value: 'Farmaceutica / Biotecnologia', label: 'Farmaceutica' },
- { value: 'Ospitalità / Hotellerie', label: 'Ospitalità' },
- { value: 'Banca / Gestione patrimoniale', label: 'Banca' },
- { value: 'Amministrazione Pubblica', label: 'Amm. Pubblica' },
- { value: 'Edilizia e tecnica', label: 'Edilizia' },
- { value: 'Istruzione e ricerca', label: 'Istruzione' },
 ];
 
 // ── Component ────────────────────────────────────────────────
