+++ components/community/jobAlertConstants.ts
@@ -1,0 +1,12 @@
+export const SECTORS = [
+ { value: 'Fintech / Blockchain', label: 'Fintech' },
+ { value: 'Tecnologia / Data Center', label: 'Tecnologia' },
+ { value: 'Consulenza', label: 'Consulenza' },
+ { value: 'Sanità / Assistenza', label: 'Sanità' },
+ { value: 'Farmaceutica / Biotecnologia', label: 'Farmaceutica' },
+ { value: 'Ospitalità / Hotellerie', label: 'Ospitalità' },
+ { value: 'Banca / Gestione patrimoniale', label: 'Banca' },
+ { value: 'Amministrazione Pubblica', label: 'Amm. Pubblica' },
+ { value: 'Edilizia e tecnica', label: 'Edilizia' },
+ { value: 'Istruzione e ricerca', label: 'Istruzione' },
+];
